+++ lib/framework/src/utilization.rs
@@ -150,8 +150,7 @@ impl Timer {
         self.ewma.update(utilization);
         let avg = self.ewma.average().unwrap_or(f64::NAN);
         debug!(utilization = %avg);
-        // TODO: support f64
-        self.utilization.set(avg as u64);
+        self.utilization.set(avg);
 
         // Reset overall statistics for the next reporting period.
         self.overall_start = self.span_start;
